requirements.txt
Babel==2.13.1
bcrypt==4.1.1
bidict==0.22.1
blinker==1.7.0
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-DebugToolbar==0.14.1
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
h11==0.14.0
idna==3.6
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
py-env==0.0.1
PyJWT==2.8.0
pymongo==4.6.1
pyotp==2.9.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.2
simple-websocket==1.0.0
six==1.16.0
urllib3==2.1.0
validators==0.22.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
